[#21][refactor]List를 list 패키지로 이동
